requirements.txt
# Core image processing
Pillow>=9.0.0
pyheif>=0.7.0
img2pdf>=0.4.0

# OCR and text processing
pytesseract>=0.3.10
pyspellchecker>=0.7.2

# AI and machine learning
openai>=1.0.0

# File and system utilities
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 